requirements.txt
alembic==1.11.1
apispec==6.3.0
blinker==1.6.2
certifi==2023.5.7
cfgv==3.3.1
click==8.1.3
coverage==7.2.6
decorator==5.1.1
distlib==0.3.6
filelock==3.12.0
flake8==6.0.0
flake8-commas==2.1.0
flake8-print==5.0.0
flake8-quotes==3.3.2
flake8-unused-arguments==0.0.13
Flask==2.3.2
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.4
flask-smorest==0.42.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
identify==2.5.24
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mccabe==0.7.0
mypy==1.3.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.1
platformdirs==3.5.1
pluggy==1.0.0
pre-commit==3.3.2
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.7.0
pytest==7.3.1
pytest-cov==4.1.0
pytest-env==0.8.1
pytest-lazy-fixture==0.6.3
python-dotenv==1.0.0
PyYAML==6.0
sentry-sdk==1.24.0
SQLAlchemy==2.0.15
typing_extensions==4.6.2
urllib3==2.0.2
virtualenv==20.23.0
webargs==8.2.0
Werkzeug==2.3.4
WTForms==3.0.1
